requirements.txt
python-telegram-bot
PyPDF2
reportlab
